restructured files to have server actions stored in lib/actions.ts instead of withing page file directly
