+++ app/dashboard/layout.tsx
@@ -1,4 +1,4 @@
-import SideNav from '@/app/ui/sidenav';
+import SideNav from '@/app/components/sidenav';
 
 export default function Layout({ children }: { children: React.ReactNode }){
