+++ app/dashboard/projects/[project_id]/items/[item_id]/versions/new/page.tsx
@@ -1,9 +1,7 @@
 import Link from "next/link";
-import { redirect } from "next/navigation";
-import { revalidatePath } from "next/cache";
-
 import { getCurrentUserId } from "@/app/lib/auth";
-import { fetchItemWithVersions, updateItemVersion } from "@/app/lib/data";
+import { fetchItemWithVersions } from "@/app/lib/data";
+import { createNewVersionAction } from "@/app/lib/actions";
 
 export default async function NewVersionPage({
   params,
@@ -27,30 +25,6 @@ export default async function NewVersionPage({
     return <div className="p-6">Item not found (or you don’t have access).</div>;
   }
 
-  async function createNewVersionAction(formData: FormData) {
-    "use server";
-
-    const userId = await getCurrentUserId();
-
-    const newVersion = String(formData.get("version_number") ?? "").trim();
-    const detailsRaw = String(formData.get("details") ?? "");
-    const details = detailsRaw.trim() ? detailsRaw.trim() : null;
-
-    // Basic validation (keep it simple for now)
-    if (!newVersion) {
-      throw new Error("Version number is required.");
-    }
-
-    await updateItemVersion(projectId, itemId, userId, newVersion, details);
-
-    // ensure item page + project dashboard refresh after mutation
-    revalidatePath(`/dashboard/projects/${projectId}`);
-    revalidatePath(`/dashboard/projects/${projectId}/items/${itemId}`);
-
-    // back to item history page
-    redirect(`/dashboard/projects/${projectId}/items/${itemId}`);
-  }
-
   return (
     <div className="p-6 max-w-xl">
       <div className="mb-4">
@@ -61,6 +35,8 @@ export default async function NewVersionPage({
       </div>
 
       <form action={createNewVersionAction} className="rounded-xl border bg-white p-4 space-y-4">
+        <input type="hidden" name="projectId" value={projectId} />
+        <input type="hidden" name="itemId" value={itemId} />
         <div>
           <label className="block text-sm font-medium">Version number</label>
           <input
